add in sass support
uses an emscripten compiled version of libsass and sassc to create a js library for sass.
note: import libraries need to be compiled in. currently only bourbon is added. to use, put '@import "bourbon/bourbon";' at the top of your sass script.

ALSO, file is extreamly large. not recommended for production.