get previews working
Previews weren't working because I forgot some hooks. Right now node calls the same code as the live preview, but it would probably be better to use the actual sass node package for this purpose. Renamed the sass library because I can see now that version names are going to be a problem.